autheticator.cpp originated.

+++ ezyMoneda/authenticator.h
@@ -1,6 +1,6 @@
 #ifndef AUTHETICATOR_H
 #define AUTHETICATOR_H
 
-bool isValidCreditCard(unsigned int cardnum);
+bool isValidCreditCard(size_t cardnum);
 
 #endif
